feat(configuration): refactor configuration handling and introduce validation and diff utilities
- Updated `AppInit.cs` to use the new `JacRed.Configuration.Schema` namespace.
- Added `AppConfigurationDiff` class for redacting sensitive information and computing configuration differences.
- Introduced `AppConfigurationValidator` for validating configuration objects and content.
- Refactored `AppConfigurationProvider` to utilize the new diff and validation utilities, improving code organization and maintainability.
- Created `ConfigSchema` to define sensitive fields and configuration structure.
