+++ Controllers/ConfigController.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using JacRed;
-using JacRed.Infrastructure.Configuration;
+using JacRed.Configuration.Schema;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
